samples: sensor: added general samples for triggers
Added samples that show off trigger functionality, removed fxos8700 driver
specific sample
