+++ boards/shields/frdm_stbc_agm01/doc/index.rst
@@ -43,8 +43,8 @@ Pin Assignment of the FRDM-STBC-AGM01 Shield
 For more information about the FXOS8700, FXAS21002, and FRDM-STBC-AGM01
 board:
 
-- :zephyr:code-sample:`fxos8700`
 - :zephyr:code-sample:`fxas21002`
+- `FXOS8700 Website`_
 - `FRDM-STBC-AGM01 Website`_
 - `FRDM-STBC-AGM01 Quick Reference Card`_
 - `FRDM-STBC-AGM01 Schematics`_
@@ -68,3 +68,6 @@ Set ``--shield frdm_stbc_agm01`` when you invoke ``west build``. For example:
 
 .. _FRDM-STBC-AGM01 Schematics:
    https://www.nxp.com/downloads/en/schematics/FRDM-STBC-AGM01-SCH.pdf
+
+.. _FXOS8700 Website:
+   https://www.nxp.com/products/sensors/accelerometers/digital-motion-sensor-3d-accelerometer-2g-4g-8g-plus-3d-magnetometer:FXOS8700CQ
