+++ samples/sensor/accel_trig/CMakeLists.txt
@@ -1,8 +1,9 @@
+# Copyright (c) 2024 NXP
 # SPDX-License-Identifier: Apache-2.0
 
 cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(fxos8700)
+project(accel_trig)
 
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
