+++ samples/sensor/accel_trig/boards/frdm_k64f.conf
@@ -1,0 +1,2 @@
+CONFIG_FXOS8700_TRIGGER_OWN_THREAD=y
+CONFIG_FXOS8700_MODE_ACCEL=y
