+++ samples/sensor/accel_trig/prj.conf
@@ -1,0 +1,3 @@
+CONFIG_STDOUT_CONSOLE=y
+CONFIG_SENSOR=y
+CONFIG_CBPRINTF_FP_SUPPORT=y
