+++ samples/sensor/accel_trig/sample.yaml
@@ -1,0 +1,14 @@
+sample:
+  name: Accelerometer trigger sample
+tests:
+  sample.sensor.accel_trig:
+    harness: console
+    tags: sensors
+    filter: dt_alias_exists("accel0")
+    harness_config:
+      type: one_line
+      regex:
+        - "^\\s*[0-9A-Za-z_,+-.]*@[0-9A-Fa-f]* \\[m\/s\\^2\\]:    \
+          \\(\\s*-?[0-9\\.]*,\\s*-?[0-9\\.]*,\\s*-?[0-9\\.]*\\)$"
+    integration_platforms:
+      - frdm_k64f                       # fxos8700
