--- samples/sensor/fxos8700/boards/mimxrt1160_evk_mimxrt1166_cm4.conf
@@ -1,2 +1,0 @@
-CONFIG_FXOS8700_TRIGGER_OWN_THREAD=n
-CONFIG_FXOS8700_TRIGGER_NONE=y
