--- samples/sensor/fxos8700/boards/mimxrt1170_evk_mimxrt1176_cm7_A.conf
@@ -1,2 +1,0 @@
-CONFIG_FXOS8700_TRIGGER_OWN_THREAD=n
-CONFIG_FXOS8700_TRIGGER_NONE=y
