--- samples/sensor/fxos8700/boards/twr_ke18f.conf
@@ -1,2 +1,0 @@
-CONFIG_FXOS8700_TRIGGER_OWN_THREAD=n
-CONFIG_FXOS8700_TRIGGER_NONE=y
