--- samples/sensor/fxos8700/overlay-motion.conf
@@ -1,1 +1,0 @@
-CONFIG_FXOS8700_MOTION=y
