--- samples/sensor/fxos8700/prj.conf
@@ -1,9 +1,0 @@
-CONFIG_STDOUT_CONSOLE=y
-CONFIG_LOG=y
-CONFIG_I2C=y
-CONFIG_SENSOR=y
-CONFIG_SENSOR_LOG_LEVEL_DBG=y
-CONFIG_FXOS8700_MODE_HYBRID=y
-CONFIG_FXOS8700_TEMP=y
-CONFIG_FXOS8700_TRIGGER_OWN_THREAD=y
-CONFIG_CBPRINTF_FP_SUPPORT=y
