--- samples/sensor/fxos8700/sample.yaml
@@ -1,41 +1,0 @@
-common:
-  harness: sensor
-  tags: sensors
-  depends_on:
-    - i2c
-    - gpio
-sample:
-  name: FXOS8700 Accelerometer/Magnetometer Sensor
-tests:
-  sample.sensor.fxos8700.hybrid:
-    platform_allow:
-      - frdm_k64f
-      - hexiwear/mk64f12
-      - warp7/mcimx7d/m4
-      - frdm_kw41z
-      - rv32m1_vega/openisa_rv32m1/ri5cy
-      - twr_ke18f
-      - lpcxpresso55s16
-      - mimxrt685_evk/mimxrt685s/cm33
-      - frdm_k22f
-      - mimxrt1024_evk
-      - mimxrt595_evk/mimxrt595s/cm33
-    integration_platforms:
-      - frdm_k64f
-    extra_configs:
-      - CONFIG_FXOS8700_PULSE=y
-      - CONFIG_FXOS8700_PULSE_INT1=y
-      - CONFIG_FXOS8700_MOTION=y
-      - CONFIG_FXOS8700_MOTION_INT1=y
-      - CONFIG_FXOS8700_MAG_VECM=y
-  sample.sensor.fxos8700.accel:
-    platform_allow:
-      - frdm_kl25z
-      - bbc_microbit
-      - lpcxpresso55s69/lpc55s69/cpu0
-      - reel_board
-      - mimxrt685_evk/mimxrt685s/cm33
-      - mimxrt595_evk/mimxrt595s/cm33
-    integration_platforms:
-      - bbc_microbit
-    extra_args: CONF_FILE=prj_accel.conf
