+++ samples/sensor/magn_trig/boards/frdm_k64f.conf
@@ -1,0 +1,1 @@
+CONFIG_FXOS8700_TRIGGER_OWN_THREAD=y
